Adding support for multiple filters. PRE and POST filters. Perhaps we need to be able to have more explicit order. State is a bit unstable right now.
Created a RequestHelper class that have some commonly used methods for when dealing with RestClientReportRequest. 
Adding task RequestReportTask that will just run a query to reporting and set result to a property. Might be used for when a filter is not ideal/worth the effort to build up q new query.